Remove unnecesary and misleading components
