+++ admin-data-explorer-slinky/src/main/scala/net/wiringbits/webapp/utils/ui/web/components/pages/DataExplorerPage.scala
@@ -1,7 +1,9 @@
 package net.wiringbits.webapp.utils.ui.web.components.pages
 
+import net.wiringbits.webapp.utils.api.models.AdminGetTablesResponse
+import net.wiringbits.webapp.utils.slinkyUtils.components.core.AsyncComponent
 import net.wiringbits.webapp.utils.ui.web.API
-import net.wiringbits.webapp.utils.ui.web.components.widgets.ExperimentalTablesWidget
+import net.wiringbits.webapp.utils.ui.web.components.widgets.{ExperimentalTableListWidget, Loader}
 import slinky.core.FunctionalComponent
 import slinky.core.annotations.react
 
@@ -9,7 +11,13 @@ import slinky.core.annotations.react
   case class Props(api: API)
 
   val component: FunctionalComponent[Props] = FunctionalComponent[Props] { props =>
-    ExperimentalTablesWidget(props.api)
+    AsyncComponent.component[AdminGetTablesResponse](
+      AsyncComponent
+        .Props(
+          fetch = () => props.api.client.getTables(),
+          render = response => ExperimentalTableListWidget.component(ExperimentalTableListWidget.Props(response)),
+          progressIndicator = () => Loader()
+        )
+    )
   }
-
 }
